+++ tests/testthat/test-googleAuthR.R
@@ -1,0 +1,5 @@
+context("Auth")
+
+test_that("Can authenticate default settings", {
+  expect_equal(2 * 2, 4)
+})